prevent redundancy when initializing oauth2server
